chore: extract GoToTopBtn logic into new component
